Start migrating special case endpoints to new pattern
