+++ server/service/handler.go
@@ -31,7 +31,6 @@ type FleetEndpoints struct {
 	CreateUserWithInvite         endpoint.Endpoint
 	PerformRequiredPasswordReset endpoint.Endpoint
 	VerifyInvite                 endpoint.Endpoint
-	EnrollAgent                  endpoint.Endpoint
 	CarveBlock                   endpoint.Endpoint
 	InitiateSSO                  endpoint.Endpoint
 	CallbackSSO                  endpoint.Endpoint
@@ -63,8 +62,6 @@ func MakeFleetServerEndpoints(svc fleet.Service, urlPrefix string, limitStore th
 		// logged in user
 		PerformRequiredPasswordReset: logged(canPerformPasswordReset(makePerformRequiredPasswordResetEndpoint(svc))),
 
-		// Osquery endpoints
-		EnrollAgent: logged(makeEnrollAgentEndpoint(svc)),
 		// For some reason osquery does not provide a node key with the block
 		// data. Instead the carve session ID should be verified in the service
 		// method.
@@ -80,7 +77,6 @@ type fleetHandlers struct {
 	CreateUserWithInvite         http.Handler
 	PerformRequiredPasswordReset http.Handler
 	VerifyInvite                 http.Handler
-	EnrollAgent                  http.Handler
 	CarveBlock                   http.Handler
 	InitiateSSO                  http.Handler
 	CallbackSSO                  http.Handler
@@ -100,7 +96,6 @@ func makeKitHandlers(e FleetEndpoints, opts []kithttp.ServerOption) *fleetHandle
 		CreateUserWithInvite:         newServer(e.CreateUserWithInvite, decodeCreateUserRequest),
 		PerformRequiredPasswordReset: newServer(e.PerformRequiredPasswordReset, decodePerformRequiredPasswordResetRequest),
 		VerifyInvite:                 newServer(e.VerifyInvite, decodeVerifyInviteRequest),
-		EnrollAgent:                  newServer(e.EnrollAgent, decodeEnrollAgentRequest),
 		CarveBlock:                   newServer(e.CarveBlock, decodeCarveBlockRequest),
 		InitiateSSO:                  newServer(e.InitiateSSO, decodeInitiateSSORequest),
 		CallbackSSO:                  newServer(e.CallbackSSO, decodeCallbackSSORequest),
@@ -288,7 +283,6 @@ func attachFleetAPIRoutes(r *mux.Router, h *fleetHandlers) {
 	r.Handle("/api/v1/fleet/sso/callback", h.CallbackSSO).Methods("POST").Name("callback_sso")
 	r.Handle("/api/v1/fleet/users", h.CreateUserWithInvite).Methods("POST").Name("create_user_with_invite")
 	r.Handle("/api/v1/fleet/invites/{token}", h.VerifyInvite).Methods("GET").Name("verify_invite")
-	r.Handle("/api/v1/osquery/enroll", h.EnrollAgent).Methods("POST").Name("enroll_agent")
 	r.Handle("/api/v1/osquery/carve/block", h.CarveBlock).Methods("POST").Name("carve_block")
 }
 
@@ -441,6 +435,13 @@ func attachNewStyleFleetAPIRoutes(r *mux.Router, svc fleet.Service, logger kitlo
 	he.POST("/api/_version_/osquery/distributed/write", submitDistributedQueryResultsEndpoint, submitDistributedQueryResultsRequestShim{})
 	he.POST("/api/_version_/osquery/carve/begin", carveBeginEndpoint, carveBeginRequest{})
 	he.POST("/api/_version_/osquery/log", submitLogsEndpoint, submitLogsRequest{})
+
+	// unauthenticated endpoints - most of those are either login-related,
+	// invite-related or host-enrolling. So they typically do some kind of
+	// one-time authentication by verifying that a valid secret token is provided
+	// with the request.
+	ne := newNoAuthEndpointer(svc, opts, r, "v1")
+	ne.POST("/api/v1/osquery/enroll", enrollAgentEndpoint, enrollAgentRequest{})
 }
 
 // TODO: this duplicates the one in makeKitHandler
