--- server/service/transport_osquery.go
@@ -1,17 +1,0 @@
-package service
-
-import (
-	"context"
-	"encoding/json"
-	"net/http"
-)
-
-func decodeEnrollAgentRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	var req enrollAgentRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		return nil, err
-	}
-	defer r.Body.Close()
-
-	return req, nil
-}
